INT-2478 put back the old canAdd logic to SimpleMessageGroup
