+++ spring-integration-core/src/test/java/org/springframework/integration/store/SimpleMessageGroupTests.java
@@ -23,18 +23,8 @@ public class SimpleMessageGroupTests {
 
 	private SimpleMessageGroup group = new SimpleMessageGroup(Collections.<Message<?>> emptyList(), key);
 
-	@SuppressWarnings("unchecked")
-	public void prepareForSequenceAwareMessageGroup() throws Exception{
-		Class<SimpleMessageGroup> clazz =
-				(Class<SimpleMessageGroup>)Class.forName("org.springframework.integration.aggregator.AbstractCorrelatingMessageHandler$SequenceAwareMessageGroup");
-		Constructor<SimpleMessageGroup> ctr = clazz.getDeclaredConstructor(MessageGroup.class);
-		ctr.setAccessible(true);
-		group = ctr.newInstance(group);
-	}
-
 	@Test
 	public void shouldFindSupersedingMessagesIfSequenceAware() throws Exception{
-		this.prepareForSequenceAwareMessageGroup();
 		final Message<?> message1 = MessageBuilder.withPayload("test").setSequenceNumber(1).build();
 		final Message<?> message2 = MessageBuilder.fromMessage(message1).setSequenceNumber(1).build();
 		assertThat(group.canAdd(message1), is(true));
@@ -45,7 +35,6 @@ public class SimpleMessageGroupTests {
 
 	@Test
 	public void shouldIgnoreMessagesWithZeroSequenceNumberIfSequenceAware() throws Exception{
-		this.prepareForSequenceAwareMessageGroup();
 		final Message<?> message1 = MessageBuilder.withPayload("test").build();
 		final Message<?> message2 = MessageBuilder.fromMessage(message1).build();
 		assertThat(group.canAdd(message1), is(true));
